+++ src/lib.rs
@@ -12,13 +12,15 @@ extern crate pancurses;
 extern crate serde;
 extern crate serde_json;
 
-mod town_level;
+mod player_action;
 mod pregame;
+mod town_level;
 
 pub mod commands;
 pub mod constants;
 pub mod conversion;
 pub mod data;
+#[macro_use]
 pub mod debug;
 pub mod equipment;
 pub mod error;