+++ src/player/data.rs
@@ -499,6 +499,7 @@ pub fn current_bulk() -> u16 {
 pub fn max_bulk() -> u16 {
     let min_base_bulk = 3000;
     let player_weight_modifier = 13;
+    // [60-280]
     let player_carry_base_amount = (30 + (player::curr_stats().strength * 10)) as u16;
     let base_bulk = min(
         (player_carry_base_amount * player_weight_modifier) + current_weight(),